requirements.txt
typer==0.12.3
rich==13.7.1
pyyaml==6.0.2
tqdm==4.66.4
tenacity==8.3.0
